refactor: export `download` from artifacts pkg for node envs
