+++ packages/artifacts/src/index.browser.ts
@@ -1,0 +1,3 @@
+export * from './index.shared'
+import maybeGetSnarkArtifacts from './download/index.browser'
+export { maybeGetSnarkArtifacts }
