+++ packages/artifacts/src/index.node.ts
@@ -1,0 +1,4 @@
+export { download } from './download/download'
+export * from './index.shared'
+import maybeGetSnarkArtifacts from './download/index.node'
+export { maybeGetSnarkArtifacts }
