+++ packages/artifacts/src/index.shared.ts
@@ -1,6 +1,3 @@
-import maybeGetSnarkArtifacts from './download/download.node'
-
 export * from './download/types'
 export { getAvailableVersions } from './download/urls'
 export * from './projects'
-export { maybeGetSnarkArtifacts }
